Moved the *Task classes into a distinct task package
