+++ src/main/java/com/ffxivcensus/gatherer/player/PlayerBuilder.java
@@ -20,7 +20,7 @@ import org.jsoup.select.Elements;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.ffxivcensus.gatherer.GathererTask;
+import com.ffxivcensus.gatherer.task.GathererTask;
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.JsonNode;
 import com.mashape.unirest.http.Unirest;
