+++ src/main/java/com/ffxivcensus/gatherer/task/GatheringLimiterTask.java
@@ -1,4 +1,4 @@
-package com.ffxivcensus.gatherer;
+package com.ffxivcensus.gatherer.task;
 
 import java.util.concurrent.ThreadPoolExecutor;
